+++ Block/PixelCode.php
@@ -12,10 +12,11 @@
 
 namespace Magestat\FacebookPixel\Block;
 
+use Magento\Framework\View\Element\Template\Context;
+use Magento\Framework\Locale\ResolverInterface;
 use Magento\Cookie\Helper\Cookie;
 use Magento\Framework\Serialize\Serializer\Json;
-use Magento\Framework\Locale\ResolverInterface;
-use Magento\Framework\View\Element\Template\Context;
+use Magento\Directory\Model\PriceCurrency;
 use Magestat\FacebookPixel\Model\PixelConfigurationInterface;
 
 /**
@@ -30,11 +31,12 @@ class PixelCode extends AbstractPixel
     private $pixelConfiguration;
 
     /**
-     * Product constructor.
+     * Checkout constructor.
      * @param Context $context
      * @param ResolverInterface $locale
      * @param Cookie $cookieHelper
      * @param Json $jsonHelper
+     * @param PriceCurrency $price
      * @param PixelConfigurationInterface $pixelConfiguration
      * @param array $data
      */
@@ -43,10 +45,11 @@ class PixelCode extends AbstractPixel
         ResolverInterface $locale,
         Cookie $cookieHelper,
         Json $jsonHelper,
+        PriceCurrency $price,
         PixelConfigurationInterface $pixelConfiguration,
         array $data
     ) {
-        parent::__construct($context, $locale, $cookieHelper, $jsonHelper, $pixelConfiguration, $data);
+        parent::__construct($context, $locale, $cookieHelper, $jsonHelper, $price, $pixelConfiguration, $data);
         $this->pixelConfiguration = $pixelConfiguration;
     }